Add slate replacement files
The PSU for the LED sign has failed, and we don't run the LEDs much since
it triggers the motion-sensitive cameras. The plastic plaques are
difficult to read without the LEDs. So, I'm working on replacing the
plastic plaques with LASER-engraved slate tiles instead. These are the
design files for that.